자동로그인 수정, 캘린더 서버 변경사항 적용 시작, 비트 기기 연결시에만 처리 적용

+++ lib/data/network/binarual_beat_recipe_response.dart
@@ -3,8 +3,8 @@ import 'base_response.dart';
 class BinauralBeatRecipeResponse extends BaseResponse {
   int state = BaseResponse.STATE_NO_CONNECT;
   String text = "";
-  double tone = 0; //Tone
-  double binauralBeat = 0; //Binaural Beat
+  int tone = 0; //Tone
+  int binauralBeat = 0; //Binaural Beat
 
   BinauralBeatRecipeResponse({
     this.state=BaseResponse.STATE_NO_CONNECT,
